\name{nea-package}
\alias{nea-package}
\alias{nea}
\docType{package}
\title{
Network Enrichment Analysis
}
\description{
Network enrichment analysis ... bla bla bla
}
\details{
\tabular{ll}{
Package: \tab nea\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2010-12-12\cr
License: \tab GLP (>=2)\cr
LazyLoad: \tab yes\cr
}
Write some text
}
\author{
Florian Markowetz

Maintainer: Florian Markowetz <florian@markowetzlab.org>
}
\references{
http://www.markowetzlab.org/software
}
\keyword{ package }
\seealso{
\code{\link[nem]{nem}}
}
\examples{
y=1
x=2
x+y
}
